Sound Fix, Reset Settings, Positioning Fix
Sound files broke because FFXIVAPP changed where they are stored, so I
have updated the default paths and appended the proper path for the
hardcoded ones.

A reset button has been added to aid in debugging.

Twintania widget can be safely moved to the top without windows trying
to maximize, and thus breaking it.